pydrake: Create `util` directory, move `type_pack` and `ModuleShim` inside.
