+++ bindings/pydrake/util/__init__.py
@@ -1,0 +1,1 @@
+# Blank Python module.
