+++ bindings/pydrake/util/module_shim.py
