+++ bindings/pydrake/util/test/module_shim_example/__init__.py
@@ -1,4 +1,4 @@
-from pydrake.util import ModuleShim
+from pydrake.util.module_shim import ModuleShim
 
 value = 1
