+++ bindings/pydrake/util/test/module_shim_test.py
